@@ -8,34 +8,21 @@ buildscript {
     }
 }
 
-apply plugin: 'groovy'
-apply plugin: 'kotlin'
+plugins {
+    id 'groovy'
+    id 'kotlin'
+    id 'com.github.dcendents.android-maven'
+    id 'com.jfrog.bintray'
+    id 'org.jetbrains.dokka'
+}
 
-sourceCompatibility = 1.7
+def VERSION_NAME = "$project.versions.testify"
+sourceCompatibility = 1.8
 
 dependencies {
     implementation "com.android.tools.build:gradle:${versions.androidGradlePlugin}"
     implementation gradleApi()
     implementation depends.kotlin.stdlib7
-}
-
-task configureLocal {
-    version = "local"
-    jar.manifest.attributes.put("Implementation-Version", version)
-}
-
-task copyLocal(type: Copy) {
-    from "build/libs/${project.name}-local.jar"
-    into 'jar'
-}
-
-task assembleLocal(description: 'Assembles the outputs of this project for use locally.', group: 'Build') {
-    dependsOn configureLocal
-    dependsOn build
-    dependsOn copyLocal
-
-    build.mustRunAfter configureLocal
-    copyLocal.mustRunAfter build
 }
 
 repositories {
