@@ -1,27 +1,43 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2019 Shopify Inc.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
 package com.shopify.testify
 
-import android.app.Activity
-import android.content.Context
-import android.graphics.Bitmap
-import android.support.test.rule.ActivityTestRule
-import android.support.test.runner.AndroidJUnit4
 import android.view.View
-
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-
-import java.io.File
-
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.rule.ActivityTestRule
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertTrue
+import org.junit.Rule
+import org.junit.Test
+import org.junit.runner.RunWith
+import java.io.File
 
 @RunWith(AndroidJUnit4::class)
 class ScreenshotUtilityTest {
 
-    @Rule
-    var testActivityRule = ActivityTestRule(TestActivity::class.java)
+    @get:Rule var testActivityRule = ActivityTestRule(TestActivity::class.java)
 
     @Test
     fun loadBaselineBitmapForComparison() {
@@ -46,21 +62,6 @@ class ScreenshotUtilityTest {
     }
 
     @Test
-    fun compareBitmaps() {
-        val activity = testActivityRule.activity
-        val rootView = activity.findViewById<View>(R.id.test_root_view)
-        val screenshotUtility = ScreenshotUtility()
-
-        val capturedBitmap = screenshotUtility.createBitmapFromActivity(activity, "testing", rootView)
-        val baselineBitmap = screenshotUtility.loadBaselineBitmapForComparison(activity, "test")
-
-        val compareDifferentBitmaps = screenshotUtility.compareBitmaps(capturedBitmap, baselineBitmap)
-        val compareSameBitmaps = screenshotUtility.compareBitmaps(baselineBitmap, baselineBitmap)
-        assertFalse(compareDifferentBitmaps)
-        assertTrue(compareSameBitmaps)
-    }
-
-    @Test
     fun deleteBitmap() {
         val activity = testActivityRule.activity
         val screenshotUtility = ScreenshotUtility()
@@ -74,5 +75,4 @@ class ScreenshotUtilityTest {
         val fileThatShouldBeDeleted = File(outputFilePath)
         assertFalse(fileThatShouldBeDeleted.exists())
     }
-
 }
