@@ -1,3 +1,26 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2019 Shopify Inc.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
 package com.shopify.testify
 
 import android.app.Activity
@@ -5,35 +28,39 @@ import android.app.Instrumentation
 import android.content.Intent
 import android.content.pm.ActivityInfo.SCREEN_ORIENTATION_LANDSCAPE
 import android.content.pm.ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
+import android.graphics.Bitmap
 import android.os.Bundle
 import android.os.Debug
 import android.os.Looper
-import android.support.annotation.FloatRange
-import android.support.annotation.IdRes
-import android.support.annotation.IntRange
-import android.support.annotation.LayoutRes
-import android.support.test.InstrumentationRegistry
-import android.support.test.InstrumentationRegistry.getInstrumentation
-import android.support.test.espresso.Espresso
-import android.support.test.rule.ActivityTestRule
 import android.util.Pair
 import android.view.View
 import android.view.ViewGroup
+import androidx.annotation.IdRes
+import androidx.annotation.LayoutRes
+import androidx.test.espresso.Espresso
+import androidx.test.platform.app.InstrumentationRegistry
+import androidx.test.platform.app.InstrumentationRegistry.getInstrumentation
+import androidx.test.rule.ActivityTestRule
 import com.shopify.testify.annotation.BitmapComparisonExactness
 import com.shopify.testify.annotation.ScreenshotInstrumentation
 import com.shopify.testify.annotation.TestifyLayout
+import com.shopify.testify.internal.DeviceIdentifier
+import com.shopify.testify.internal.compare.FuzzyCompare
+import com.shopify.testify.internal.compare.SameAsCompare
 import com.shopify.testify.internal.exception.AssertSameMustBeLastException
 import com.shopify.testify.internal.exception.NoScreenshotsOnUiThreadException
 import com.shopify.testify.internal.exception.RootViewNotFoundException
 import com.shopify.testify.internal.exception.ScreenshotBaselineNotDefinedException
 import com.shopify.testify.internal.exception.ScreenshotIsDifferentException
+import com.shopify.testify.internal.helpers.FontScaleHelper
+import com.shopify.testify.internal.helpers.LocaleHelper
 import com.shopify.testify.internal.modification.HideCursorViewModification
 import com.shopify.testify.internal.modification.HidePasswordViewModification
 import com.shopify.testify.internal.modification.HideScrollbarsViewModification
 import com.shopify.testify.internal.modification.HideTextSuggestionsViewModification
 import com.shopify.testify.internal.modification.SoftwareRenderViewModification
-import junit.framework.Assert.assertNotNull
-import junit.framework.Assert.assertTrue
+import org.junit.Assert.assertNotNull
+import org.junit.Assert.assertTrue
 import org.junit.rules.TestRule
 import org.junit.runner.Description
 import org.junit.runners.model.Statement
@@ -41,66 +68,69 @@ import java.util.Locale
 import java.util.concurrent.CountDownLatch
 import java.util.concurrent.TimeUnit
 
-
 typealias ViewModification = (rootView: ViewGroup) -> Unit
 typealias EspressoActions = () -> Unit
 typealias ViewProvider = (rootView: ViewGroup) -> View
+typealias BitmapCompare = (baselineBitmap: Bitmap, currentBitmap: Bitmap) -> Boolean
 
-open class ScreenshotRule<T : Activity> @JvmOverloads constructor(private val activityClass: Class<T>, initialTouchMode: Boolean = false, launchActivity: Boolean = true)
-    : ActivityTestRule<T>(activityClass, initialTouchMode, launchActivity), TestRule {
-
-    @LayoutRes private var targetLayoutId: Int = NO_ID
-    @IdRes private var rootViewId: Int = android.R.id.content
-
-    private var bitmapCompare: BitmapCompare? = null
-    private var defaultLocale: Locale? = null
-    private var locale: Locale? = null
-    private var exactness: Float? = null
-
-    private var assertSameInvoked = false
-    private var hideSoftKeyboard = true
-    private var hideScrollbars = true
-    private var hidePasswords = true
-    private var hideCursor = true
-    private var hideTextSuggestions = true
-    private var isLayoutInspectionModeEnabled = false
-    private var useSoftwareRenderer = false
-
-    lateinit var testMethodName: String
-    private lateinit var testSimpleClassName: String
-    private lateinit var testClass: String
-    private var activityMonitor: Instrumentation.ActivityMonitor? = null
-
-    private var viewModification: ViewModification? = null
-    private var espressoActions: EspressoActions? = null
-    private var screenshotViewProvider: ViewProvider? = null
-    private val testContext = InstrumentationRegistry.getContext()
-
-    private val hidePasswordViewModification = HidePasswordViewModification()
-    private val hideScrollbarsViewModification = HideScrollbarsViewModification()
-    private val hideTextSuggestionsViewModification = HideTextSuggestionsViewModification()
-    private val softwareRenderViewModification = SoftwareRenderViewModification()
-    private val hideCursorViewModification = HideCursorViewModification()
-
-    private var throwable: Throwable? = null
+@Suppress("unused")
+open class ScreenshotRule<T : Activity> @JvmOverloads constructor(
+    private val activityClass: Class<T>,
+    initialTouchMode: Boolean = false,
+    launchActivity: Boolean = true
+) : ActivityTestRule<T>(activityClass, initialTouchMode, launchActivity), TestRule {
 
     constructor(activityClass: Class<T>, @IdRes rootViewId: Int) : this(activityClass) {
         this.rootViewId = rootViewId
     }
 
+    @IdRes private var rootViewId: Int = android.R.id.content
+    @LayoutRes private var targetLayoutId: Int = NO_ID
+    @Suppress("MemberVisibilityCanBePrivate")
+    lateinit var testMethodName: String
+    private lateinit var testClass: String
+    private lateinit var testSimpleClassName: String
+    private val hideCursorViewModification = HideCursorViewModification()
+    private val hidePasswordViewModification = HidePasswordViewModification()
+    private val hideScrollbarsViewModification = HideScrollbarsViewModification()
+    private val hideTextSuggestionsViewModification = HideTextSuggestionsViewModification()
+    private val softwareRenderViewModification = SoftwareRenderViewModification()
+    private val testContext = getInstrumentation().context
+    private var activityMonitor: Instrumentation.ActivityMonitor? = null
+    private var assertSameInvoked = false
+    private var defaultFontScale: Float? = null
+    private var defaultLocale: Locale? = null
+    private var espressoActions: EspressoActions? = null
+    private var exactness: Float? = null
+    private var fontScale: Float? = null
+    private var hideCursor = true
+    private var hidePasswords = true
+    private var hideScrollbars = true
+    private var hideSoftKeyboard = true
+    private var hideTextSuggestions = true
+    private var isLayoutInspectionModeEnabled = false
+    private var locale: Locale? = null
+    private var screenshotViewProvider: ViewProvider? = null
+    private var throwable: Throwable? = null
+    private var useSoftwareRenderer = false
+    private var viewModification: ViewModification? = null
+
+    @Suppress("MemberVisibilityCanBePrivate")
     val testName: String
-        get() = testSimpleClassName + "_" + testMethodName
+        get() = "${testSimpleClassName}_$testMethodName"
 
-    private val testNameComponents: Boolean
-        get () = Looper.getMainLooper().thread == Thread.currentThread()
+    private fun isRunningOnUiThread(): Boolean {
+        return Looper.getMainLooper().thread == Thread.currentThread()
+    }
 
-    private val getTestNameComponents : Pair<String, String>
+    private val testNameComponents: Pair<String, String>
         get() = Pair(testSimpleClassName, testMethodName)
 
-    private val fullyQualifiedTestPath : String
+    private val fullyQualifiedTestPath: String
         get() = testClass
 
-    fun setExactness(@FloatRange(from = 0.0, to = 1.0) exactness: Float): ScreenshotRule<T> {
+    fun setExactness(exactness: Float): ScreenshotRule<T> {
+        require(exactness in 0.0..1.0)
         this.exactness = exactness
         return this
     }
@@ -166,6 +196,11 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(private val ac
         return this
     }
 
+    fun setFontScale(fontScale: Float): ScreenshotRule<T> {
+        this.fontScale = fontScale
+        return this
+    }
+
     fun setScreenshotViewProvider(viewProvider: ViewProvider): ScreenshotRule<T> {
         this.screenshotViewProvider = viewProvider
         return this
@@ -182,7 +217,8 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(private val ac
      *
      * @param requestedOrientation SCREEN_ORIENTATION_LANDSCAPE or SCREEN_ORIENTATION_PORTRAIT
      */
-    fun setOrientation(@IntRange(from = SCREEN_ORIENTATION_LANDSCAPE.toLong(), to = SCREEN_ORIENTATION_PORTRAIT.toLong()) requestedOrientation: Int) {
+    fun setOrientation(requestedOrientation: Int): ScreenshotRule<T> {
+        require(requestedOrientation in SCREEN_ORIENTATION_LANDSCAPE..SCREEN_ORIENTATION_PORTRAIT)
         if (activity.requestedOrientation != requestedOrientation) {
             val activityMonitor = getActivityMonitor()
             getInstrumentation().addMonitor(activityMonitor)
@@ -190,6 +226,7 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(private val ac
             activity.requestedOrientation = requestedOrientation
             InstrumentationRegistry.getInstrumentation().waitForIdleSync()
         }
+        return this
     }
 
     private fun getActivityMonitor(): Instrumentation.ActivityMonitor {
@@ -213,14 +250,12 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(private val ac
 
     override fun apply(base: Statement, description: Description): Statement {
         checkForScreenshotInstrumentationAnnotation(description)
-
         applyExactness(description)
-
         espressoActions = null
         testSimpleClassName = description.testClass.simpleName
         testMethodName = description.methodName
-        testClass = description.testClass.canonicalName + "#" + description.methodName
-        val testifyLayout = description.getAnnotation(TestifyLayout::class.java)
+        testClass = "${description.testClass?.canonicalName}#${description.methodName}"
+        val testifyLayout: TestifyLayout? = description.getAnnotation(TestifyLayout::class.java)
         targetLayoutId = testifyLayout?.layoutId ?: View.NO_ID
         return super.apply(ScreenshotStatement(base), description)
     }
@@ -235,7 +270,6 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(private val ac
         }
     }
 
-    // todo merge getActivityIntent() and getIntent()
     override fun getActivityIntent(): Intent {
         var intent: Intent? = super.getActivityIntent()
         if (intent == null) {
@@ -244,6 +278,7 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(private val ac
         return intent
     }
 
+    @Suppress("MemberVisibilityCanBePrivate")
     protected fun getIntent(): Intent {
         var intent = super.getActivityIntent()
         if (intent == null) {
@@ -259,11 +294,6 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(private val ac
         }
     }
 
-    /**
-     * @param exactness 0.0: matches against anything(no differences)
-     *                  1.0: exact matches only
-     * @throws Exception same as assertSame()
-     */
     fun assertSame() {
         assertSameInvoked = true
         try {
@@ -272,12 +302,17 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(private val ac
                 LocaleHelper.setTestLocale(it)
             }
 
-            exactness?.let {
-                bitmapCompare = FuzzyCompare(it)
+            fontScale?.let {
+                defaultFontScale = testContext.resources.configuration.fontScale
+                FontScaleHelper.setTestFontScale(it)
             }
 
+            val bitmapCompare: BitmapCompare = exactness?.let {
+                FuzzyCompare(it)::compareBitmaps
+            } ?: SameAsCompare()::compareBitmaps
+
             try {
-                if (testNameComponents) {
+                if (isRunningOnUiThread()) {
                     throw NoScreenshotsOnUiThreadException()
                 }
 
@@ -301,7 +336,7 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(private val ac
                     screenshotView = screenshotViewProvider!!.invoke(getRootView(activity))
                 }
 
-                val outputFileName = DeviceIdentifier.formatDeviceString(DeviceIdentifier.DeviceStringFormatter(testContext, getTestNameComponents), getOutputFileNameFormatString())
+                val outputFileName = DeviceIdentifier.formatDeviceString(DeviceIdentifier.DeviceStringFormatter(testContext, testNameComponents), getOutputFileNameFormatString())
                 val currentBitmap = screenshotUtility.createBitmapFromActivity(activity, outputFileName, screenshotView)
                 assertNotNull("Failed to capture bitmap from activity", currentBitmap)
 
@@ -317,11 +352,7 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(private val ac
                         throw ScreenshotBaselineNotDefinedException(getModuleName(), testName, fullyQualifiedTestPath)
                     }
 
-                if (bitmapCompare == null) {
-                    bitmapCompare = screenshotUtility
-                }
-
-                if (bitmapCompare!!.compareBitmaps(baselineBitmap, currentBitmap)) {
+                if (bitmapCompare(baselineBitmap, currentBitmap!!)) {
                     assertTrue("Could not delete cached bitmap $testName", screenshotUtility.deleteBitmap(activity, outputFileName))
                 } else {
                     if (isRecordMode()) {
@@ -333,6 +364,9 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(private val ac
             } finally {
                 defaultLocale?.let {
                     LocaleHelper.setTestLocale(it)
+                }
+                defaultFontScale?.let {
+                    FontScaleHelper.setTestFontScale(it)
                 }
             }
         } finally {
@@ -403,7 +437,7 @@ open class ScreenshotRule<T : Activity> @JvmOverloads constructor(private val ac
         val extras = InstrumentationRegistry.getArguments()
         var formatString: String = DeviceIdentifier.DEFAULT_NAME_FORMAT
         if (extras.containsKey("outputFileNameFormat")) {
-            formatString = extras.getString("outputFileNameFormat")
+            formatString = extras.getString("outputFileNameFormat")!!
         }
         return formatString
     }
