@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2017 Shopify Inc.
+ * Copyright (c) 2019 Shopify Inc.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,6 +24,7 @@
 
 package com.shopify.testify.internal.exception
 
-class ScreenshotIsDifferentException(moduleName: String, testName: String) : Exception("\n\n*  The captured screenshot is different from the baseline screenshot.\n" +
-    "*  Run `./gradlew " + moduleName + ":screenshotPull` to view the differences.\n" +
-    "*  Run `./gradlew " + moduleName + ":screenshotTest -PtestClass=" + testName + "` to run this test again.\n\n")
+class ScreenshotIsDifferentException(moduleName: String, testName: String) : Exception(
+    "\n\n*  The captured screenshot is different from the baseline screenshot.\n" +
+            "*  Run `./gradlew $moduleName:screenshotPull` to view the differences.\n" +
+            "*  Run `./gradlew $moduleName:screenshotTest -PtestClass=$testName` to run this test again.\n\n")
