@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2017 Shopify Inc.
+ * Copyright (c) 2019 Shopify Inc.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -26,13 +26,7 @@ package com.shopify.testify.internal.modification
 import android.view.View
 import android.view.ViewGroup
 
-abstract class ViewModification(enabled: Boolean) {
-
-    var isEnabled = false
-
-    init {
-        this.isEnabled = enabled
-    }
+abstract class ViewModification(private val isEnabled: Boolean = false) {
 
     fun modify(view: View) {
         if (!isEnabled) {
